avt-command: Ignore unmatched noun parts instead of erroring
When parsing the second part of a noun, if it doesn’t match the
properties of the first part then ignore it instead of making the whole
noun parsing fail. That way if it can be added to a different part of
speech then it will work. For example in the sentence “al la plaĝo hundo
iros”, if “hundo” is ignored when parsing “plaĝo” then it can become a
standalone subject.