avt-state: Implement the $D delay command for rule messages
When a delay is encountered in a message the rest of the message is now
sent as a DELAY type. play-avt ignores this and just joins it with the
previous message. The JavaScript code starts a new span for the message
part and displays it after a timeout with a fadein animation.