avt-state: Make a seperate function to ensure carrying movable
Instead of having a different function to ensure we are carrying the
movable for each position in the command, the function is now split out
into a helper function that just ensures whatever movable we need is
being carried. The callers then need to do the check in two parts.