The “nothing” action is actually supposed to remove the object
The documentation for the compiler says that if you choose “nenio” then
the object “estas for”. At first I didn’t understand this but in
practice it seems to mean the object disappears. This is a little weird
because it is the default action and nothing in the documentation
mentions a way to do nothing. However, specifying “io” for the action
does seem to do nothing so that provides a way to do it.