avt: Make the list of rule conditions and actions unlimited
Instead of having a single slot for each type of condition and action
like in the original data format, the conditions and actions are now
stored as an array where each entry contains the slot number. This paves
the way to add more slot types later and makes it possible to internally
represent rules with multiple conditions and actions of the same type.