avt: Allow multiple verbs per rule
When the rules are used by the state machine they are only ever
searched by verb name. That means we don’t really know the verb for each
rule. Instead it is better to have each verb maintain a list of each
rule that uses it. This should potentially make it faster to scan for
rules corresponding to a verb. It also makes it easy to have a rule with
multiple verbs. This patch makes the parser allow that.