parser: Add a new command to replace the object
The commands that replace an object in the original interpreter are
pretty confusing and it’s still not really clear what they are supposed
to do. It seems like the most useful thing to do would be to just
remove the old object and make the new object appear in the same place
as the old one, whether that be carried or in the room. This patch adds
a new command that does that and makes the “nova aĵo <x>” action use
that instead of the old actions.