ludu-avt: Improve decoding the editor DOM for Chrome
Chrome handles the DOM slightly differently than Firefox because the
first line isn’t in a separate div. This patch makes it insert a new
line at the start of the div instead of at the end, and only if it isn’t
already at the start of a line. That way it doesn’t need to have any
special handling of the br tags at the end of the element as well.