ludu-avt: Scroll to the top of new messages
Previously whenever a new message was added the message area would be
scrolled to the bottom of the messages. If there is a message that is
longer than the screen the player would then have to scroll back up to
read from the start of the message. To fix this the messages are now
added to a group which is scrolled to using scrollIntoView whenever a
new message is added. That way if it is too long then it will scroll to
the top of the group instead. A new group is started whenever a command
is typed.